_technique/requirements.txt
weasyprint>=60.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.32.0
mistralai>=1.0.0
jsonschema>=4.20.0
pytest>=8.0.0
ruff>=0.5.0
